Fix initialization.  Add failed attempt at pseudo parsing.
I was hoping to be clever by bypassing parsing and examining data
directly, but this is not possible - the only way to get at the
packet data is in a parser function, via extract or set_metadata.
We will have to parse but it will be messy because we need to
selectively parse based on switch state, but the only way to get
at the switch state is by applying a table, and we can only apply
tables after parsing is done.  Therefore, our only recourse to
overcome this catch-22 is to do the following when each packet
arrives: 1) check that local metadata has not been populated with
persistent switch state and skip parsing, sending control to the
main control function; 2) apply a table that populates local
metadata with persistent switch state; 3) resubmit and return
to parsing.