Implement draft modify_field functionality.  Add test framework.
HyPer4 is thought to have functionality for one mode of the modify_field
function.  A testing framework has been added with supporting scripts.
I cannot currently execute the tests because for some reason I am having
trouble connecting to the CLI, and not just for HyPer4, but for demo
targets such as l2_switch and basic_router in p4factory/targets.