Demo bmv1 bug(hp4test1). Create header obfuscation and demo.
The hp4test1.p4 program exposed a bug in bmv1 that was fixed today.

Header obfuscation is a fallback project given that HyPer4 progress
is so uncertain.  It demonstrates a concept described by the HyPer4
paper, in which the bytes in a header are shuffled.  The first
effort is very simple and not yet dynamic without human intervention.
A complete demo is included in this commit, with instructions.