Restructure repo.  Fix header obfuscation.  Two demos included.
Most of the work in this commit is represented by a fixed
header obfuscation program; the previous commit included a flawed
attempt that not only did not carry out the intended permutation,
but did not take into account the parallel semantics for
primitives invoked in a P4 compound action.  The version in this
commit fixes the permutation and significantly reduces (but does
not eliminate) the issues arising from parallel execution of
primitives in compound actions.  A complete solution will require
breaking the obfsl2 compound action across two tables.