Hone in on p4c-bmv2/hlir hangup.
Identified cause of useless conditional nodes being removed - I had
empty control functions in add_header.p4 awaiting implementation.
Unfortunately, those were not the cause of p4c-bmv2/hlir hanging
when trying to produce hp4.json.  I suspect table dependency
checks are the cause.  This commit includes test2, in which hp4
is being reassembled in the simplest form possible, periodically
producing a JSON in the hope that I can spot the point at which
p4c-bmv2/hlir hangs.  This commit's version of test2 currently
does not produce the hang symptom even though it represents a
substantial chunk of hp4.  But I suspect the full hp4 is simply
too big for p4c-bmv2/hlir to handle; if that is true, hp4 will
have to be put on hold until p4c-bmv2/hlir is upgraded.