Add compile script and brief analysis of complexity in terms of tables.
Contrary to suspicions in previous commits, p4c-bmv2 does not hang
when we run it on hp4, but it takes a long time, which time appears
to increase exponentially in the number of tables involved.

The compile script reports real, user, and sys time required, and
will send an SMS when complete if a cell number is supplied as the
first argument.

Analysis shows the number of hp4 tables grows linearly with the
number of matching stages supportable and number of primitives
per stage supported, as 4 + a*(9 + 6*b), where a = num stages
and b = num primitives per stage.

I've managed to compile a version of hp4 with 2 stages and 3
primitives per stage, requiring 25 minutes.  I must inquire
about skipping table dependency checks.