Fix and improve multicast, add hub target.
Multicast now uses the concept of a sequence id.  This essentially
is a multicast group id.  The operator can define custom sequences
of ports for packet replication.

The hub target is a concise demonstration of this technique of
multicasting.  The technique developed in this target was grafted
into HyPer4, but hub is also a suitable objective for HyPer4
demonstration.