Add header_test, demo'ing the use of header stacks in place of metadata.
Metadata has a limitation in that we cannot make arrays out of it.
Where we desire this, header stacks may be used.  All header instances
not meant to be serialized with the rest of the packet during
deparsing must be removed from the parsed representation at some
point.  Probably the egress pipeline is a good place to do it.