Add pushpop_test.  Demo: bmv2 doesn't support modify_field(dest, val, mask).
This is disappointing.  Despite modify_field(dest, val, mask) still included
in the latest, yet-to-be-official language spec, attempts to use it result
in a compiler warning that it is no longer supported in bmv2, and that the
code is automatically replaced with modify_field(dest, val & mask), which
is not equivalent.  Sent an e-mail asking why.