Add test projects to deepen understanding of simple_router.
Simple_router demo allows pings to flow even when dropping anything
that doesn't include an ipv4 header.  This is because ARP tables
are set in the mininet scenario.  We don't even need to rewrite
the source MAC when forwarding packets.

As yet it is unclear why HyPer4-simple_router isn't working, but
I've narrowed the possibilities.  Will examine packet traces next.