Simple_router: implement manual ipv4 checksum; successful (non-HP4) test.
The 'P4 way' of doing checksums is not changeable at runtime - it uses
something called a 'field_list_calculation.'  The bottom line is it cannot
be compiled into the table operations HyPer4 requires to emulate a program
that needs checksums.

So the neat thing about this commit is it paves the way for a general
method of doing runtime-definable checksums.  Next step: graft into
HyPer4 and proceed with HyPer4-simple_router testing.