Complete demo of two programs running within HyPer4 SIMULTANEOUSLY!
This commit includes a script that starts mininet with six hosts,
all connected to one switch.  Hosts 1-3 are on the same subnet,
hosts 4-6 are all on different subnets.  The switch runs
HyPer4 and the script feeds it a series of commands that configures
it to emulate a L2 switch program (hp4t_l2_switch) for any
traffic on ports 1-3, and a L3 routing program (simple_router) for
any traffic on ports 4-6.  Tables for each program are also
populated as required.  Hosts 1-3 can all ping each other, as can
hosts 4-6.