Fix normalization bug, complete simple_router demo conversion.
We now have two working demos!  After completing the changes to
the commands file for the simple_router demo, I discovered that
the normalizer function a_norm_20_39 was missing some parenthesis
and mucking up the packet.  And but so now both the L2 switch
and the L3 router demos work with the new version of HP4 using
the flexible parser.