Continue implementation of packet structure manipulation primitives.
See especially add_header_notes.txt for a detailed discussion of
how add_header should work.  I daresay we are ready to punch in
the code and give things a go.  The header_manip_test project
is intended for trying concepts out prior to integration with
the rest of HP4.