Add more test projects exploring bmv2 limitations regarding the spec.
In a series of disappointing exchanges with the P4 developers, I
have learned that in the cases I've discovered via test projects
in this commit where bmv2 does not adhere to the language spec,
the response is that the language spec needs to be revised to reflect
implementation realities.  Meaning bmv2 won't be fixed; the standard
will.  Meaning many of the possible ways to have HP4 support packet
structure manipulation are closed off.

The original idea for implementing add_header in HP4 required obscene
amounts of boilerplate, mostly auto-generated code.  The alternative
was more complex logically but ideally should have required one-tenth
the amount of code.  The reality for this alternative approach is
the code savings will not be nearly as significant as hoped; it
remains to be seen whether the final code savings justify the approach.