Partial implementation of add_header + code generation script.
This is partial because while it changes the parsed representation,
it does not change the metadata field extracted.data that is
actually used during the ingress pipeline.

Also, I believe there is yet another approach that may simplify
packet structure manipulation drastically, and has the potential
to render the approach found in this commit obsolete.

The idea is to track how many bytes should be in the parsed
representation if all packet structure manipulation primitives
were applied correctly, and delay modifying ext[] until the
end, right before copying extracted.data to the elements of
ext.  The modification of ext simply consists of one pop and/or
push based on the difference between how large it was right
after parsing and the value in the field tracking the new
size.