Fix multicast bug in smtag projects; likely need to apply to hp4 too.
This bug only showed up in a multi-switch demo: a packet enters
port 1 and requires clones of the packet be sent out all other
ports, but lo, a clone is also sent back out port 1, and with
multiple switches carrying ethernet broadcast frames they start
to multiply exponentially, saturating links, spiking CPU utlization,
and basically threatening to bring down all of civilization as
maybe a Fifth Horseman of the Apocalypse.

But it's fixed now.