Demo that sswitch can clone AND recirculate the same packet.
The success of clone_recirc_test is a relief because HP4 multicasting
relies on this behavior.  After designing HP4 multicast, I was
worried that if a packet was marked for cloning and recirculation,
sswitch would only do one of the two at the end of egress, but
fortunately it does both.