Change add_to_field to more general math_on_field.  Add subtraction support.
This is a small expansion of HyPer4 ability to support math on fields; we've
simply added support for subtracting a constant.  But this points the way
toward general mathematical expression support, which I previously thought
was impossible.  There is nothing stopping us from implementing separate
methods for each mathematical operation, e.g., +,-,*,/,%,<<,>>,&,| and
then we can string them together to support expressions (support for
parenthesis would be implemented in the compiler).