Stop auto generating pcaps in mininet scripts.  Delete iface_down.sh.
Pcaps can still be generated, if desired, by supplying --pcap to
the topo* mininet script.  But they are no longer generated
automatically.

The iface_down.sh script is obsolete now that the topo* scripts
issue nmcli dev disconnect commands as of the previous commit.
Therefore it is deleted in this commit.