Overhaul HP4 code generation for easier development.
Bottom line: we can write P4 code augmented with a few special
markers to aid automatic generation of functionally redundant
(but name-differentiated) final output P4 code.

This is a significant improvement over the previous method that
employed python scripts to produce HP4 P4 code.  When examining
these scripts it was so difficult to "see" the P4 code, that
development typically consisted of 1) producing working .P4
code, 2) revising the python scripts as necessary, and 3) comparing
the output of the python scripts with the hand-crafted .p4 files,
returning to 2) as necessary.  This process was slow and error-
prone.

The HP4 code generation method now employs what is essentially a
DSL for any code that has much functional redundancy that should
be expressable with loop constructs.  For example, the P4 compiler
does not allow multiple references to control functions or tables,
and because HP4 needs to provide functionality at arbitrary points
in the packet processing pipeline, HP4 requires functionally
redundant control functions and tables be defined that differ
only in name.  The DSL in this commit allows the following
constructs:
- [+ sloop [# num iterations] ["singlespace"|"ss"] +]
  [code including the marker "[+X+]" -- this marker is replaced
   with the loop iteration]
  [+ endsloop +]
  Typically, sloop is used without specifying the number of
  iterations, in which case this parameter is taken from the
  command line argument --numstages/-ns.
  Unless the singlespace option is provided, an empty line is
  inserted between output sections of code.
- [+ dloop [# num outer iterations] [# num inner iterations] +]
  [code including markers "[+X+]" and "[+Y+]" -- these markers
   are replaced with the outer and inner loop iterations]
  [+ enddloop +]
  dloop is like sloop but employs a double loop.  Unless otherwise
  specified, the number of outer iterations is the command line
  argument --numstages/-ns, and the number of inner iterations is
  taken from --numprimitives/-np.
- [+ nif [# num iterations] +]
  [code including marker "[+Y+]" -- this marker is replaced with
   the loop iteration]
  [+ endnif +]
  nif stands for Nested If.  It is similar to sloop but increases
  the indent for the code block output with each iteration, and
  takes care of the closing braces as well.  nif assumes the first
  line of code includes an opening brace.
- [+ std_header +]
  Inserts the contents of data/stdheader_d.  C-style comment
  delimiters are added to the output.
- [+ this_header +]
  Inserts the contents of a file, the name of which is computed
  by starting with /data, concatenating the name of the .p4t being
  processed but without the .p4t extension, and finally concatenating
  "_d".  C-style comment delimiters are added to the output.

Note that comment handling has been added such that if std_header
or this_header is followed directly by this_header or std_header,
the output is a single comment (i.e., only one set of C-style
delimiters appears in the output).