Refactor and optimize HP4 (in terms of table lookups).
This commit implements changes proposed in optimization_tables_notes.txt,
except for required changes to demos (specifically, the commands
files).  For these demos, some of the changes will be search-and-replace
to change table names, but the changes to the compound action emulation
system require more effort due to table and action consolidation.