Generate tset_control_state commands for simplest cases.
Every piece of this parse tree handling is a morass of logic,
lots of corner cases.  Right now the puzzle is identifying the
right range of bytes to inspect when a parse node has a select
expression in the return statement.  It varies due to the
possibility of 'current' expressions that look into the packet
without extracting anything.  There could possibly be two
paths that arrive at a parse node with the same number of bits
having been extracted, but in one of these paths, some of those
bits were extracted in order to handle 'current' expressions,
while in the other path there were no such 'current' expressions,
and so the offsets pertaining to headers and 'current' expressions
in the parse node would be different for each path.